Login implemented

+++ project.py
@@ -1,8 +1,21 @@
-from flask import Flask, render_template, request, redirect, url_for
+from flask import (Flask, render_template, request, redirect, url_for,
+                   make_response, flash)
+app = Flask(__name__)
 from sqlalchemy import create_engine, asc, desc
 from sqlalchemy.orm import sessionmaker
-from database_setup import Base, Category, Item
-app = Flask(__name__)
+from database_setup import Base, Category, Item, User
+from flask import session as login_session
+import random
+import string
+import json
+import httplib2
+import requests
+from oauth2client.client import flow_from_clientsecrets, FlowExchangeError
+
+
+CLIENT_ID = json.loads(
+    open('client_secrets.json', 'r').read())['web']['client_id']
+APPLICATION_NAME = "Restaurant Menu Application"
 
 
 engine = create_engine('sqlite:///catalog.db')
@@ -12,32 +25,204 @@ DBSession = sessionmaker(bind=engine)
 session = DBSession()
 
 
+def create_user(login_session):
+    new_user = User(name=login_session['username'], email=login_session[
+                   'email'], picture=login_session['picture'])
+    session.add(new_user)
+    session.commit()
+    user = session.query(User).filter_by(email=login_session['email']).one()
+    return user.id
+
+
+def get_user_info(user_id):
+    user = session.query(User).filter_by(id=user_id).one()
+    return user
+
+
+def get_user_ID(email):
+    try:
+        user = session.query(User).filter_by(email=email).one()
+        return user.id
+    except Exception:
+        return None
+
+
+def get_user_from_session(login_session):
+    if 'email' in login_session:
+        return get_user_info(get_user_ID(login_session['email']))
+    return None
+
+
 @app.route("/")
 def list_latest():
     categories = session.query(Category).order_by(asc(Category.name))
     title_items = 'Latest Items'
     items = session.query(Item).order_by(desc(Item.time_created))
+    user = get_user_from_session(login_session)
     return render_template('categories.html', categories=categories,
-                           title_items=title_items, items=items)
+                           title_items=title_items, items=items, user=user)
+
+
+@app.route('/login')
+def show_login():
+    state = ''.join(random.choice(string.ascii_uppercase + string.digits)
+                    for x in xrange(32))
+    login_session['state'] = state
+    return render_template('login.html', STATE=state)
+
+
+@app.route('/gconnect', methods=['POST'])
+def gconnect():
+    # Validate state token
+    if request.args.get('state') != login_session['state']:
+        response = make_response(json.dumps('Invalid state parameter.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    # Obtain authorization code
+    code = request.data
+
+    try:
+        # Upgrade the authorization code into a credentials object
+        oauth_flow = flow_from_clientsecrets('client_secrets.json', scope='')
+        oauth_flow.redirect_uri = 'postmessage'
+        credentials = oauth_flow.step2_exchange(code)
+    except FlowExchangeError:
+        response = make_response(
+            json.dumps('Failed to upgrade the authorization code.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Check that the access token is valid.
+    access_token = credentials.access_token
+    url = ('https://www.googleapis.com/oauth2/v1/tokeninfo?access_token=%s'
+           % access_token)
+    h = httplib2.Http()
+    result = json.loads(h.request(url, 'GET')[1])
+    # If there was an error in the access token info, abort.
+    if result.get('error') is not None:
+        response = make_response(json.dumps(result.get('error')), 500)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is used for the intended user.
+    gplus_id = credentials.id_token['sub']
+    if result['user_id'] != gplus_id:
+        response = make_response(
+            json.dumps("Token's user ID doesn't match given user ID."), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Verify that the access token is valid for this app.
+    if result['issued_to'] != CLIENT_ID:
+        response = make_response(
+            json.dumps("Token's client ID does not match app's."), 401)
+        print "Token's client ID does not match app's."
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    stored_access_token = login_session.get('access_token')
+    stored_gplus_id = login_session.get('gplus_id')
+    if stored_access_token is not None and gplus_id == stored_gplus_id:
+        response = make_response(json.dumps('Current user is already' +
+                                            ' connected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
+    # Store the access token in the session for later use.
+    login_session['access_token'] = credentials.access_token
+    login_session['gplus_id'] = gplus_id
+
+    # Get user info
+    userinfo_url = "https://www.googleapis.com/oauth2/v1/userinfo"
+    params = {'access_token': credentials.access_token, 'alt': 'json'}
+    answer = requests.get(userinfo_url, params=params)
+
+    data = answer.json()
+
+    login_session['username'] = data['name']
+    login_session['picture'] = data['picture']
+    login_session['email'] = data['email']
+
+    user_id = get_user_ID(login_session['email'])
+    if not user_id:
+        user_id = create_user(login_session)
+    login_session['user_id'] = user_id
+
+    output = ''
+    output += '<h1>Welcome, '
+    output += login_session['username']
+    output += '!</h1>'
+    output += '<img src="'
+    output += login_session['picture']
+    output += (' " style = "width: 300px; height: 300px;border-radius:' +
+               ' 150px;-webkit-border-radius: 150px;-moz-border-radius:' +
+               ' 150px;"> ')
+    flash("you are now logged in as %s" % login_session['username'])
+    print "done!"
+    return output
+
+
+@app.route('/gdisconnect')
+def gdisconnect():
+    access_token = login_session['access_token']
+    print 'In gdisconnect access token is %s', access_token
+    print 'User name is: '
+    print login_session['username']
+    if access_token is None:
+        print 'Access Token is None'
+        response = make_response(json.dumps(
+            'Current user not connected.'), 401)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    url = 'https://accounts.google.com/o/oauth2/revoke?token=%s' % login_session[
+        'access_token']
+    h = httplib2.Http()
+    result = h.request(url, 'GET')[0]
+    print 'result is '
+    print result
+    if result['status'] == '200':
+        del login_session['access_token']
+        del login_session['gplus_id']
+        del login_session['username']
+        del login_session['email']
+        del login_session['picture']
+        del login_session['user_id']
+        response = make_response(json.dumps('Successfully disconnected.'), 200)
+        response.headers['Content-Type'] = 'application/json'
+        return response
+    else:
+
+        response = make_response(json.dumps(
+            'Failed to revoke token for given user.', 400))
+        response.headers['Content-Type'] = 'application/json'
+        return response
+
 
 @app.route("/new", methods=['GET', 'POST'])
 def new_item():
+    if 'username' not in login_session:
+        return redirect('/login')
     if request.method == 'GET':
+        user = get_user_from_session(login_session)
         categories = session.query(Category).order_by(asc(Category.name))
-        return render_template('newitem.html', categories=categories)
+        return render_template('newitem.html', categories=categories,
+                               user=user)
 
     if request.method == 'POST':
         item = Item(name=request.form['name'],
                     description=request.form['description'],
-                    category_id=int(request.form['category']))
+                    category_id=int(request.form['category']),
+                    user_id=login_session['user_id'])
         session.add(item)
         session.commit()
         return redirect(url_for('list_latest'))
 
+
 @app.route("/catalog/item/<int:item_id>")
 def item_details(item_id):
     item = session.query(Item).filter_by(id=item_id).one()
-    return render_template('itemdetails.html', item=item)
+    user = get_user_from_session(login_session)
+    return render_template('itemdetails.html', item=item, user=user)
 
 
 @app.route("/catalog/<int:category_id>")
@@ -45,15 +230,21 @@ def items_category(category_id):
     categories = session.query(Category).order_by(asc(Category.name))
     items = session.query(Item).filter_by(category_id=category_id)
     category = session.query(Category).filter_by(id=category_id).one()
+    user = get_user_from_session(login_session)
     return render_template('categories.html', title_items=category.name,
-                           items=items, categories=categories)
+                           items=items, categories=categories, user=user)
 
 
 @app.route("/delete/item/<int:item_id>", methods=['GET', 'POST'])
 def delete_item(item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
+    if login_session['email'] != item.user.email:
+        return 'You cannot delete a item that isn\'t yours '
     if request.method == 'GET':
-        return render_template('deleteitem.html', item=item)
+        user = get_user_from_session(login_session)
+        return render_template('deleteitem.html', item=item, user=user)
     if request.method == 'POST':
         session.delete(item)
         session.commit()
@@ -62,11 +253,16 @@ def delete_item(item_id):
 
 @app.route("/edit/item/<int:item_id>", methods=['GET', 'POST'])
 def edit_item(item_id):
+    if 'username' not in login_session:
+        return redirect('/login')
     item = session.query(Item).filter_by(id=item_id).one()
+    if login_session['email'] != item.user.email:
+        return 'You cannot edit a item that isn\'t yours '
     if request.method == 'GET':
+        user = get_user_from_session(login_session)
         categories = session.query(Category).order_by(asc(Category.name))
         return render_template('edititem.html', item=item,
-                               categories=categories)
+                               categories=categories, user=user)
     if request.method == 'POST':
         item.name = request.form['name']
         item.description = request.form['description']
@@ -77,5 +273,6 @@ def edit_item(item_id):
 
 
 if __name__ == "__main__":
+    app.secret_key = 'SUPER_SECRET_KEY'
     app.debug = True
     app.run(host='0.0.0.0', port=5000)
